Re-enabled the unit tests for RabbitInAHat; merged the 'hide tables' test into the existing test for RabbitInAHat
