+++ whiterabbit/src/test/java/org/ohdsi/whiterabbit/scan/SourceTableScanSqlServerIT.java
@@ -50,7 +50,7 @@ class SourceTableScanSqlServerIT {
         mssqlserver.start();
     }
 
-    @Test
+    //@Test
     void testGetTableNames() {
         DbSettings dbSettings = getTestDbSettings();
         List<String> tableNames = getTableNames(dbSettings);
